Layout of WRAP Temperature Sensor Type B
For printing on paper using Dimatix 2831
Type B: Automated switching, NTC modulates amplitude
Usage: Automated switching (with astable multivibrator) for temperature data collection. The amplitude difference between ON and OFF period is related to the sensed temperature. The layout is shown for a length of 13 cm, which can be changed by simply extending the arm with the temperature sensor (NTC).