Layout of WRAP Temperature Sensor Type C
For printing on paper using Dimatix 2831
Type C: Automated switching, NTC modulates time period of OFF time
Usage: Automated switching (with astable multivibrator) for temperature data collection. The time duration of ON period is constant for reference. The time duration of OFF period is related to the sensed temperature. The duty cycle can also be related to the sensed temperature. The layout is shown for a length of 13 cm, which can be changed by simply extending the arm with the temperature sensor (NTC).