Layout of WRAP Temperature Sensor Type D
For printing on paper using Dimatix 2831
Type D: Single shot switching, NTC modulates time period of switching
Usage: Automated single-shot switching (with RC delay) for temperature data collection. The time duration of switching is related to the sensed temperature. The depth of amplitude before and after switching can be related to mutual inductance. The layout is shown for a length of 13 cm, which can be changed by simply extending the arm with the temperature sensor (NTC).